refactor(activities): move relink-only helpers out of lifecycle.go
relinkedChangedFields is RelinkActivity's own event-field builder, called
only from relinkbatch.go where relinkAdmittedRow/relinkActivityRow live —
lifecycle.go's held-row fix pushed it to 501 lines, past craft static's
500-line cap. Pure code motion, no behavior change.
